Switch to nanosleep(2) from usleep(2)
The usleep(2) function was deprecated by POSIX.1-2001 and removed in
POSIX.1-2008. It was replaced by nanosleep(2), which was first defined
in POSIX.1b.

Further, getopt(3) was defined in POSIX.2 and sigaction(2) in POSIX.1.
In summary, instead of _DEFAULT_SOURCE, you this program really only
needs _POSIX_C_SOURCE at least 199309L.

Except that, unfortunately, SIGWINCH is not defined by POSIX, though it
will next in the next standard:

http://austingroupbugs.net/view.php?id=1151